Viewer course, wave 1: src ordered by lesson, docs/cut.py writes every crate from lessons/37, raw chain re-cut (write-once 0)
- src: later-lesson code moved into trailing per-lesson sections, extracted hooks and tagged
  register lines; engine/gpu/slots.rs and app/walk/plane.rs split out. Same behaviour.
- docs/lessons: 43 crates cut from the master 37 (new 18a, 18b, 23a-23d; 27, 28, 29, 34 folded
  in and their pages removed); FILES.txt and REGISTER.txt give each file and tag its lesson.
- check_write_once 0 (was 12973), check_lesson37 0, cut.py --check 0, all crates check lib + native.
- Pages still name pre-cut step anchors: mkdocs.yml check_paths is false until they are rewritten,
  and the Vue course plugin honours the same switch (warns instead of failing).

+++ session_viewer/docs/lessons/10/src/shaders/grid.wgsl
@@ -15,8 +15,8 @@ const BLUE: vec3<f32> = vec3<f32>(0.30, 0.45, 0.85);
 
 // One grid vertex.
 struct VsOut {
-    @builtin(position) pos: vec4<f32>,
-    @location(0) color: vec3<f32>,
+    @builtin(position) pos: vec4<f32>, // clip position
+    @location(0) color: vec3<f32>, // line color
 }
 
 @vertex
@@ -62,5 +62,5 @@ fn vs_main(@builtin(vertex_index) vid: u32) -> VsOut {
 @fragment
 // Flat line color.
 fn fs_main(in: VsOut) -> PhysicalColor {
-    return PhysicalColor(vec4<f32>(in.color, 1.0), vec2<f32>(0.0));
+    return PhysicalColor(vec4<f32>(in.color, 1.0), vec2<u32>(0u));
 }
